src/requirements.txt
openai>=1.0.0
PyGithub>=2.1.1
tree-sitter
